+++ doc/conf.py
@@ -27,6 +27,7 @@ import sasmodels
 
 nitpick_ignore = [
     ('py:class', 'argparse.Namespace'),
+    ('py:class', 'bumps.parameter.Parameter'),
     ('py:class', 'collections.OrderedDict'),
     ('py:class', 'cuda.Context'),
     ('py:class', 'cuda.Function'),
@@ -40,6 +41,7 @@ nitpick_ignore = [
     ('py:class', 'pyopencl._cl.Device'),
     ('py:class', 'pyopencl._cl.Kernel'),
     ('py:class', 'QWebView'),
+    ('py:class', 'types.ModuleType'),
     ('py:class', 'unittest.suite.TestSuite'),
     ('py:class', 'wx.Frame'),
     # autodoc and namedtuple is completely broken
@@ -55,6 +57,7 @@ nitpick_ignore = [
     ('py:class', 'module'),
     ('py:class', 'SesansData'),
     ('py:class', 'SourceModule'),
+    ('py:class', 'TestCondition'),
     # KernelModel and Calculator breaking on git actions tests, even though
     # KernelModel is already listed. astropy example sometimes includes full
     # path to complaining symbol. Let's see if that helps here:
@@ -140,40 +143,6 @@ pygments_style = 'sphinx'
 
 # A list of ignored prefixes for module index sorting.
 #modindex_common_prefix = []
-
-nitpick_ignore = [
-    ('py:class', 'argparse.Namespace'),
-    ('py:class', 'bumps.parameter.Parameter'),
-    ('py:class', 'collections.OrderedDict'),
-    ('py:class', 'cuda.Context'),
-    ('py:class', 'cuda.Function'),
-    ('py:class', 'np.dtype'),
-    ('py:class', 'numpy.dtype'),
-    ('py:class', 'np.ndarray'),
-    ('py:class', 'numpy.ndarray'),
-    ('py:class', 'pyopencl.Program'),
-    ('py:class', 'pyopencl._cl.Context'),
-    ('py:class', 'pyopencl._cl.CommandQueue'),
-    ('py:class', 'pyopencl._cl.Device'),
-    ('py:class', 'pyopencl._cl.Kernel'),
-    ('py:class', 'QWebView'),
-    ('py:class', 'unittest.suite.TestSuite'),
-    ('py:class', 'wx.Frame'),
-    # autodoc and namedtuple is completely broken
-    ('py:class', 'integer -- return number of occurrences of value'),
-    ('py:class', 'integer -- return first index of value.'),
-    # autodoc doesn't handle these type definitions
-    ('py:class', 'Data'),
-    ('py:class', 'Data1D'),
-    ('py:class', 'Data2D'),
-    ('py:class', 'Kernel'),
-    ('py:class', 'ModelInfo'),
-    ('py:class', 'module'),
-    ('py:class', 'SesansData'),
-    ('py:class', 'SourceModule'),
-    ('py:class', 'TestCondition'),
-]
-
 
 
 # -- Options for HTML output ---------------------------------------------------